修改了InBrand接口,的添加接口里面的itemId的类型应该为String.因为前端使用nanoid库去添加itemID,对出库记录的校验也是根据itemId是否相同进行校验.

+++ src/main/java/com/bsqx/bsqx_rear/service/InBound/InboundServicesImp.java
@@ -4,21 +4,10 @@ import com.bsqx.bsqx_rear.DTO.Inbound.Inbound;
 
 import java.util.List;
 
-/**
- * # -*- coding: utf-8 -*-
- *
- * @Time ： 2024/4/16 16:24
- * @Author ： SuYan
- * @File ：InboundServicesImp.java
- * @email: suyan1254088@gmail.com
- * @IDE ：IntelliJ IDEA 2021.2.2
- * @Motto ：ABC(Always Be Coding)
- */
 public interface InboundServicesImp {
-
     Inbound addInBound(Inbound inbound);
-    Inbound deleteInBound(Long id);
-    Inbound editInBound(int id,Inbound inbound);
+    Inbound deleteInBound(Integer id);
+    Inbound editInBound(int id, Inbound inbound);
     List<Inbound> searchInBound(String keyWord);
     List<Inbound> allInbound();
 }
